requirements.txt
# Web Framework
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Session==0.5.0

# Machine Learning & Data Science
numpy>=1.22,<2.0  # TensorFlow 2.x requires numpy < 2.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Enhanced Multimodal
 (TensorFlow)
tensorflow>=2.13.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Google AI Integration
google-generativeai==0.3.2

# Security & Authentication
Werkzeug==2.3.7

# Environment Variables
python-dotenv==1.0.0

# Data Processing
openpyxl==3.1.2

# Production Server
gunicorn==21.2.0

# Database
pymongo==4.5.0

# Additional utilities
pathlib2==2.3.7

# Voice Analysis
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.16.0